% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kr.R
\name{kr_spread}
\alias{kr_spread}
\title{Recent spread data}
\usage{
kr_spread(pair, since = NULL)
}
\arguments{
\item{pair}{Trading pair}

\item{since}{Unix timestamp to start from (optional)}
}
\value{
tibble: timestamp, bid, ask, pair
}
\description{
Recent spread data
}
