% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kr.R
\name{kr_ticker}
\alias{kr_ticker}
\title{Current ticker data for one or more pairs}
\usage{
kr_ticker(pair)
}
\arguments{
\item{pair}{Trading pair(s). Single string or comma-separated.
Examples: "XBTUSD" (BTC/USD), "ETHUSD", "SOLUSD", "XBTUSD,ETHUSD"}
}
\value{
tibble: pair, ask, bid, last, volume_24h, vwap_24h, trades_24h,
low_24h, high_24h
}
\description{
Current ticker data for one or more pairs
}
