% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kr.R
\name{kr_trades}
\alias{kr_trades}
\title{Recent trades}
\usage{
kr_trades(pair, count = 100, since = NULL)
}
\arguments{
\item{pair}{Trading pair}

\item{count}{Number of trades (default 100)}

\item{since}{Trade ID to start from (optional)}
}
\value{
tibble: price, volume, timestamp, side (buy/sell), type (market/limit),
misc, pair
}
\description{
Recent trades
}
