% Generated by build script
\name{kr_ticker}
\alias{kr_ticker}
\title{Current ticker data for one or more pairs}
\usage{
kr_ticker()
}
\arguments{
\item{pair}{Trading pair(s). Single string or comma-separated.}
}
\value{tibble: pair, ask, bid, last, volume_24h, vwap_24h, trades_24h,}
\description{Current ticker data for one or more pairs}
